chore: modularized logs component
